+++ narwhals/_compliant/expr.py
@@ -914,19 +914,6 @@ class LazyExpr(  # type: ignore[misc]
     gather_every: not_implemented = not_implemented()
     replace_strict: not_implemented = not_implemented()
 
-    @property
-    def _backend_version(self) -> tuple[int, ...]:
-        return self._implementation._backend_version()
-
-    def alias(self, name: str) -> Self:
-        def fn(names: Sequence[str]) -> Sequence[str]:
-            if len(names) != 1:
-                msg = f"Expected function with single output, found output names: {names}"
-                raise ValueError(msg)
-            return [name]
-
-        return self._with_alias_output_names(fn)
-
     def _with_alias_output_names(self, func: AliasNames | None, /) -> Self: ...
 
     @property
